requirements.txt
pymysql
coc.py
asyncio
cryptography
waitress
flask
